Update README.md
Environment Variables
If you look at config/default.json, you'll see a property called jwtPrivateKey. This key is used to encrypt JSON web tokens. So, for security reasons, it should not be checked into the source control. I've set a default value here to make it easier for you to get up and running with this project. For a production scenario, you should store this key as an environment variable.

On Mac:

export vidly_jwtPrivateKey=yourSecureKey

On Windows:

set vidly_jwtPrivateKey=yourSecureKey